requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Data / ML
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2

# OCR / vision
easyocr==1.7.1
opencv-python==4.10.0.84
Pillow==10.4.0

# NLP / language
spacy>=3.7,<3.8
pyspellchecker==0.8.1

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.2

# Face similarity (optional, CPU-only)
face-recognition==1.3.0

# HTTP client
httpx==0.27.2

# Config
python-dotenv==1.0.1
pydantic-settings==2.5.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
